requirements.txt
customtkinter
pyserial
pyautogui
pycryptodome
pystray
pillow
pyperclip
